Data.Base16.Types

Description

This module contains the Base16 type definition, Alphabet datatype, alphabet constraints, and various quality of life combinators for working with Base16-wrapped data.

Documentation

data Base16 a

Wraps a value, asserting that it is or is intended to be in a particular kind of Base16 encoding use extractBase16 to extract the value, and assertBase16 to tag a value as base16-encoded

assertBase16 :: a -> Base16 a

Assert the provenance of a value.

Warning: This is a blind assertion that a particular value is base16 encoded in some alphabet. If you are not sure of the provenance of the value, you may experience odd behavior when attempting to decode. Use at your own risk. If I see any issues logged on this project from negligent use of this, Sofia and I will smite you.

extractBase16 :: Base16 a -> a

Forget that a particular value is Base16-encoded