Data.ByteString.Short.Base16

Description

This module contains ShortByteString-valued combinators for implementing the RFC 4648 specification of the Base16 encoding format. This includes lenient decoding variants, as well as internal and external validation for canonicity.

Documentation

encodeBase16 :: ShortByteString -> ShortText

Encode a ShortByteString value as Base16 ShortText with padding.

See: RFC-4648 section 8

Examples:

>>> encodeBase16 "Sun"
"53756e"

Since: 0.3.0.0

encodeBase16' :: ShortByteString -> ShortByteString

Encode a ShortByteString value as a Base16 ShortByteString value with padding.

See: RFC-4648 section 8

Examples:

>>> encodeBase16' "Sun"
"53756e"

decodeBase16 :: ShortByteString -> Either Text ShortByteString

Decode a Base16-encoded ShortByteString value.

See: RFC-4648 section 8

Examples:

>>> decodeBase16 "53756e"
Right "Sun"

>>> decodeBase16 "6x"
Left "invalid character at offset: 1"

decodeBase16' :: ShortText -> Either Text ShortByteString

Decode Base16 Text.

See: RFC-4648 section 8

Examples:

>>> decodeBase16' "53756e"
Right "Sun"

>>> decodeBase16' "6x"
Left "invalid character at offset: 1"

decodeBase16Lenient :: ShortByteString -> ShortByteString

Decode a Base16-encoded ShortByteString value leniently, using a strategy that never fails

N.B.: this is not RFC 4648-compliant

Examples:

>>> decodeBase16Lenient "53756e"
"Sun"

>>> decodeBase16Lenient "6x6x"
"f"

isBase16 :: ShortByteString -> Bool

Tell whether a ShortByteString value is base16 encoded.

Examples:

>>> isBase16 "666f6"
False

>>> isBase16 "666f"
True

isValidBase16 :: ShortByteString -> Bool

Tell whether a ShortByteString value is a valid Base16 format.

This will not tell you whether or not this is a correct Base16 representation, only that it conforms to the correct alphabet. To check whether it is a true Base16 encoded ShortByteString value, use isBase16.

Examples:

>>> isValidBase16 "666f+/6"
False

>>> isValidBase16 "666f6"
True